[Feature][JsonSplit] modify updateProcessDefinition/switchProcessDefinitionVersion  (#4666)
* processdefinition create/delete method

* init

* add relation parse

* delete process_definition_json

* modify updateProcessDefinition/switchProcessDefinitionVersion

* codestyle

* codestyle

* fix compile error （version type）

* fix compile error （version type）

+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/service/ProcessDefinitionVersionService.java
@@ -31,7 +31,7 @@ public interface ProcessDefinitionVersionService {
      * @param processDefinition the process definition that need to record version
      * @return the newest version number of this process definition
      */
-    long addProcessDefinitionVersion(ProcessDefinition processDefinition);
+    int addProcessDefinitionVersion(ProcessDefinition processDefinition);
 
     /**
      * query the pagination versions info by one certain process definition id
